Refactor and extract BaseMap component
